Revert "refactor: #554 응답의 소수 데이터를 Float 타입으로 변경 (#559)"
This reverts commit ba1dd28fe785717fa53bd8133626555975ff7322.

+++ backend/ddang/src/main/java/com/ddang/ddang/review/application/dto/CreateReviewDto.java
@@ -14,7 +14,7 @@ public record CreateReviewDto(Long auctionId, Long writerId, Long targetId, Stri
                 writerId,
                 createReviewRequest.targetId(),
                 createReviewRequest.content(),
-                createReviewRequest.score().doubleValue()
+                createReviewRequest.score()
         );
     }
